유저 정보 변경 API 구현

+++ src/main/java/com/todayhouse/infra/S3Storage/service/FileServiceImpl.java
@@ -86,11 +86,13 @@ public class FileServiceImpl implements FileService {
 
     @Override
     public void deleteOne(String fileName) {
+        if (fileName.isEmpty())
+            return;
         amazonS3.deleteObject(bucketName, fileName);
     }
 
     @Override
-    public String changeFileNameToUrl(String fileName){
+    public String changeFileNameToUrl(String fileName) {
         return amazonS3.getUrl(bucketName, fileName).toString();
     }
 }
